refactor stratified classification script
still need to fix parity in CV folds between different seeds
