+++ 02_run_stratified_classification.py
@@ -12,13 +12,14 @@ import pandas as pd
 from tqdm import tqdm
 
 import pancancer_evaluation.config as cfg
-import pancancer_evaluation.utilities.data_utilities as du
-from pancancer_evaluation.mutation_prediction import (
-    MutationPrediction,
+from pancancer_evaluation.data_models.tcga_data_model import TCGADataModel
+from pancancer_evaluation.exceptions import (
     NoTestSamplesError,
     OneClassError,
     ResultsFileExistsError
 )
+import pancancer_evaluation.utilities.data_utilities as du
+from pancancer_evaluation.utilities.classify_utilities import run_cv_stratified
 
 def process_args():
     p = argparse.ArgumentParser()
@@ -84,13 +85,13 @@ if __name__ == '__main__':
 
     sample_info_df = du.load_sample_info(args.verbose)
 
-    predictor = MutationPrediction(seed=args.seed,
-                                   results_dir=args.results_dir,
-                                   subset_mad_genes=args.subset_mad_genes,
-                                   verbose=args.verbose,
-                                   debug=args.debug)
+    tcga_data = TCGADataModel(seed=args.seed,
+                              results_dir=args.results_dir,
+                              subset_mad_genes=args.subset_mad_genes,
+                              verbose=args.verbose,
+                              debug=args.debug)
 
-    genes_df = predictor.load_gene_set(args.gene_set)
+    genes_df = tcga_data.load_gene_set(args.gene_set)
 
     # we want to run mutation prediction experiments:
     # - for true labels and shuffled labels
@@ -112,14 +113,17 @@ if __name__ == '__main__':
             outer_progress.set_description('gene: {}'.format(gene))
 
             try:
-                predictor.process_data_for_gene(gene, classification,
+                tcga_data.process_data_for_gene(gene, classification,
                                                 use_pancancer=True,
                                                 check_gene_file=True,
                                                 shuffle_labels=shuffle_labels)
             except ResultsFileExistsError:
+                # this happens if cross-validation for this gene has already been
+                # run (i.e. the results file already exists)
                 if args.verbose:
                     print('Skipping because results file exists already: gene {}'.format(
                         gene), file=sys.stderr)
+                # TODO make below into a function
                 cancer_type_log_df = pd.DataFrame(
                     dict(zip(log_columns,
                              [gene, True, shuffle_labels, 'file_exists']
@@ -131,13 +135,13 @@ if __name__ == '__main__':
                 continue
             except KeyError:
                 # this might happen if the given gene isn't in the mutation data
-                # (or has a different alias, TODO check for this later)
+                # (or has a different alias, TODO we could check for this later)
                 print('Gene {} not found in mutation data, skipping'.format(gene),
                       file=sys.stderr)
                 continue
 
             try:
-                predictor.run_cv_stratified(gene, sample_info_df,
+                results = run_cv_stratified(tcga_data, gene, sample_info_df,
                                             args.num_folds, shuffle_labels)
             except NoTestSamplesError:
                 if args.verbose:
@@ -159,6 +163,13 @@ if __name__ == '__main__':
                          )),
                     index=[0]
                 )
+            else:
+                # only save results if no exceptions
+                du.save_results_stratified(tcga_data.gene_dir,
+                                           tcga_data.check_file,
+                                           results,
+                                           gene,
+                                           shuffle_labels)
 
             if cancer_type_log_df is not None:
                 cancer_type_log_df.to_csv(args.log_file, mode='a', sep='\t',
