+++ pancancer_evaluation/utilities/data_utilities.py
@@ -4,6 +4,7 @@ Functions for reading/writing/processing data
 """
 import os
 import sys
+from pathlib import Path
 
 import numpy as np
 import pandas as pd
@@ -325,43 +326,35 @@ def summarize_results(results, gene, holdout_cancer_type, signal, z_dim,
 
     return metrics_out_, roc_df_, pr_df_
 
-def subset_by_mad(X_train_df, X_test_df, gene_features, subset_mad_genes, verbose=False):
-    """Subset features by mean absolute deviation.
 
-    Takes the top subset_mad_genes genes (sorted in descending order),
-    calculated on the training set.
+def save_results_stratified(gene_dir, check_file, results, gene, signal):
+    gene_auc_df = pd.concat(results['gene_auc'])
+    gene_aupr_df = pd.concat(results['gene_aupr'])
+    gene_coef_df = pd.concat(results['gene_coef'])
+    gene_metrics_df = pd.concat(results['gene_metrics'])
 
-    Arguments
-    ---------
-    X_train_df: training data, samples x genes
-    X_test_df: test data, samples x genes
-    gene_features: numpy bool array, indicating which features are genes (and should be subsetted/standardized)
-    subset_mad_genes (int): number of genes to take
-
-    Returns
-    -------
-    (train_df, test_df, gene_features) datasets with filtered features
-    """
-    if verbose:
-        print('Taking subset of gene features', file=sys.stderr)
-
-    mad_genes_df = (
-        X_train_df.loc[:, gene_features]
-                  .mad(axis=0)
-                  .sort_values(ascending=False)
-                  .reset_index()
+    gene_coef_df.to_csv(
+        check_file, sep="\t", index=False, compression="gzip",
+        float_format="%.5g"
     )
-    mad_genes_df.columns = ['gene_id', 'mean_absolute_deviation']
-    mad_genes = mad_genes_df.iloc[:subset_mad_genes, :].gene_id.astype(str).values
 
-    non_gene_features = X_train_df.columns.values[~gene_features]
-    valid_features = np.concatenate((mad_genes, non_gene_features))
+    output_file = Path(
+        gene_dir, "{}_{}_auc_threshold_metrics.tsv.gz".format(
+            gene, signal)).resolve()
+    gene_auc_df.to_csv(
+        output_file, sep="\t", index=False, compression="gzip", float_format="%.5g"
+    )
 
-    gene_features = np.concatenate((
-        np.ones(mad_genes.shape[0]).astype('bool'),
-        np.zeros(non_gene_features.shape[0]).astype('bool')
-    ))
-    train_df = X_train_df.reindex(valid_features, axis='columns')
-    test_df = X_test_df.reindex(valid_features, axis='columns')
-    return train_df, test_df, gene_features
+    output_file = Path(
+        gene_dir, "{}_{}_aupr_threshold_metrics.tsv.gz".format(
+            gene, signal)).resolve()
+    gene_aupr_df.to_csv(
+        output_file, sep="\t", index=False, compression="gzip", float_format="%.5g"
+    )
+
+    output_file = Path(gene_dir, "{}_{}_classify_metrics.tsv.gz".format(
+        gene, signal)).resolve()
+    gene_metrics_df.to_csv(
+        output_file, sep="\t", index=False, compression="gzip", float_format="%.5g"
+    )
